Move exported types in a separated file
